Data.H3.Scalable

Description

Type classes for scales. A scale is a family of functions that are parameterised by a target range (for example, the dimensions of the diagram in pixels) and an additional set of options specific to each scale.

In h3, Scalable is the type class that represents scales. Scales used for visualisation commonly have metadata such as legends, grid lines etc. The ChartVisuals class deals with this kind of data.

Scales

class Scalable (f :: * -> *) a b where

The class of scales. Each scale has an associated data type ScaleOptions that can be used to configure the scale.

Minimal complete definition

scale

Associated Types

type Target f :: * -> *

Target f is the codomain of the scale. For example, if f is a scale that maps real numbers to real numbers, then Target f is Identity (a point). If it is an ordinal scale, then Target f is Extent (an interval).

type TargetRange f b :: *

TargetRange f is the type of the range parameter of the scale. For example, if f is a scale that maps real numbers to real numbers, then TargetRange f is Extent (an interval). For ordinal scales, Target f is [], a list of possible values.

data ScaleOptions f a b :: *

Additional parameters for the scale (other than target range).

Methods

scale :: ScaleOptions f a b -> TargetRange f b -> a -> Target f b

Given ScaleOptions and a TargetRange, produce a map 'a -> (Target f) b'.

Basic instances

arrow :: (a -> b) -> ScaleOptions ((->) a) a b

Every function 'f :: a -> b' is a scale