Data.H3.Scales

Description

A set of useful scales for visualising data. Each scale is identified by an empty data type (Continuous, Ordinal etc). To use a scale f to map as to bs you need to construct a value of ScaleOptions f a b. To make this easier, a constructor such as continuous, ordinal, etc. is provided for each scale.

Scales can be combined in a number of ways, just like to ordinary functions. For example, we can take the Product of any two scales f and g. The ScaleOptions of combined scales are usually functions of the ScaleOptions of their components, and can be constructed in the same way.

Basic scales

arrow :: (a -> b) -> ScaleOptions ((->) a) a b

Every function 'f :: a -> b' is a scale

continuous

Arguments

:: Extent a
The source range

-> IncludeZeroPolicy
Whether to extend the source range to include 0

-> ScaleOptions Continuous a a

Create a continuous scale (map between real numbers).

data Continuous a

data IncludeZeroPolicy

Indicates whether to extend the source (domain) of a continuous scale to include 0.

Extending the scale to 0 ensures that absolute values are comparable but may obscure differences between values if the variation of the data is small compared to its absolute size.

Constructors

IncludeZero

DontIncludeZero

cardinal :: Extent a -> ScaleOptions Cardinal a Double

Create a cardinal scale (map from integers to Double).

data Cardinal a

ordinal

Arguments

:: NonEmpty a
The domain

-> (a -> String)
A description of the values (for creating labels)

-> ScaleOptions Ordinal a Double

Create an ordinal scale (map from some type with an Ord instance to 'Extent Double').

data Ordinal a

Combinators

product :: ScaleOptions f a b -> ScaleOptions g c d -> ScaleOptions (Product f g) (a, c) (b, d)

The product of two scales.

newtype Product f g a

Constructors

Product

Fields

getProduct :: (f (LeftV a), g (RightV a))

nested :: ScaleOptions f a b -> ScaleOptions g c b -> ScaleOptions (Nested f g) (a, c) b

nested takes two scales g and f and creates a g scale within each f result. This requires 'Target f', Target g and TargetRange g to be identical.

data Nested (f :: * -> *) (g :: * -> *) p

transformed :: (Target f b -> Target f b) -> ScaleOptions f a b -> ScaleOptions (Transformed f) a b

Map the results of a scale monomorphically.

data Transformed (f :: * -> *) a

anchored :: a -> ScaleOptions f a b -> ScaleOptions (Anchored f) a b

Extend a scale by "anchoring" its results to a fixed point of the domain

data Anchored (f :: * -> *) a

newtype Pair a

Constructors

Pair

Fields

getPair :: (a, a)

class ProductV p

A type that is the result of a binary function

Data family instance constructors

data family ScaleOptions f a b :: *

Additional parameters for the scale (other than target range).