Data.H3.Shape

Description

Shapes that make up charts.

Documentation

data Shape s n

Shapes that make up charts. s is the type of colours, and n the type of coordinates. For a 2D chart n should be something like (Double, Double).

Constructors

ALine [n]
An open line of several points

ARectangle !(Extent n)
A rectangle, defined by its lower left and upper right corners (if n ~ (Double, Double))

AnArea [n]
An area (closed line)

AColouredShape s !(Shape s n)
Set the colour of a shape

AGroup [Shape s n]
Group several shapes together

ALabel !TextAnchor !(FontSize String) !(FontWeight String) !String n
A label

AnOpacity Double !(Shape s n)
Set the opacity of a shape

EmptyShape
Empty shape

mapColour :: (s -> t) -> Shape s n -> Shape t n

Change the colours of a Shape.

data TextAnchor

Indicates where a text element should be anchored, that is where its origin is placed relative to the text.

Constructors

AnchorStart

AnchorMiddle

AnchorEnd

newtype FontSize s

Font size

Constructors

FontSize

Fields

getFontSize :: s

_FontSize :: forall p f a b. (Profunctor p, Functor f) => p (FontSize a) (f (FontSize b)) -> p a (f b)

Font size iso

newtype FontWeight s

Font weight

Constructors

FontWeight

Fields

getFontWeight :: s

_FontWeight :: forall p f a b. (Profunctor p, Functor f) => p (FontWeight a) (f (FontWeight b)) -> p a (f b)

Font weight iso

newtype Pixel a

Pixel unit in the target coordinate system

Constructors

Pixel

Fields

getPixel :: a

_Pixel :: forall p f a b. (Profunctor p, Functor f) => p (Pixel a) (f (Pixel b)) -> p a (f b)

Pixel iso

newtype Width

Width

Constructors

Width

Fields

getWidth :: Double

_Width :: forall p f. (Profunctor p, Functor f) => p Width (f Width) -> p Double (f Double)

Width iso

newtype Height

Height

Constructors

Height

Fields

getHeight :: Double