Data.H3.Utils

Documentation

looseLabels :: (Ord a, Num a, Fractional a, Floating a, RealFrac a) => LabelCount -> (a, a) -> (NonEmpty a, Integer)

Generate a list of "nice" numbers that contain the given interval.

Based on "Nice numbers for label graphs" by Paul Heckbert. http://www.realtimerendering.com/resources/GraphicsGems/gems/Label.c

newtype LabelCount

How many labels to generate

Constructors

LabelCount Integer

defaultLabelCount :: LabelCount

data RoundingMode

Constructors

Round

DontRound

nicenum :: (Ord a, Num a, Fractional a, Floating a, RealFrac a) => RoundingMode -> a -> a

linear :: (Fractional a, Eq a) => (a, a) -> (a, a) -> a -> a

computeMidpoint :: (Fractional x, Num x, Foldable1 f) => f x -> x