Data.H3.Visuals

Description

Shapes that make up charts.

Shapes

data Shape s n

Shapes that make up charts. s is the type of colours, and n the type of coordinates. For a 2D chart n should be something like (Double, Double).

Constructors

ALine [n]
An open line of several points

ARectangle !(Extent n)
A rectangle, defined by its lower left and upper right corners (if n ~ (Double, Double))

AnArea [n]
An area (closed line)

AColouredShape s !(Shape s n)
Set the colour of a shape

AGroup [Shape s n]
Group several shapes together

ALabel !TextAnchor !(FontSize String) !(FontWeight String) !String n
A label

AnOpacity Double !(Shape s n)
Set the opacity of a shape

EmptyShape
Empty shape

mapColour :: (s -> t) -> Shape s n -> Shape t n

Change the colours of a Shape.

data TextAnchor

Indicates where a text element should be anchored, that is where its origin is placed relative to the text.

Constructors

AnchorStart

AnchorMiddle

AnchorEnd

newtype FontSize s

Font size

Constructors

FontSize

Fields

getFontSize :: s

_FontSize :: forall p f a b. (Profunctor p, Functor f) => p (FontSize a) (f (FontSize b)) -> p a (f b)

Font size iso

newtype FontWeight s

Font weight

Constructors

FontWeight

Fields

getFontWeight :: s

_FontWeight :: forall p f a b. (Profunctor p, Functor f) => p (FontWeight a) (f (FontWeight b)) -> p a (f b)

Font weight iso

newtype Pixel a

Pixel unit in the target coordinate system

Constructors

Pixel

Fields

getPixel :: a

_Pixel :: forall p f a b. (Profunctor p, Functor f) => p (Pixel a) (f (Pixel b)) -> p a (f b)

Pixel iso

newtype LabelOffset a

Constructors

LabelOffset

Fields

getLabelOffset :: a

_LabelOffset :: forall p f a b. (Profunctor p, Functor f) => p (LabelOffset a) (f (LabelOffset b)) -> p a (f b)

LabelOffset iso

data VisualElements s n

A collection of scale metadata that needs to be visualised. n is the type of coordinates in the target coordinate system, to allow combinators such as Product and Cartesian to adjust the locations of the elements.

Constructors

VisualElements

Fields

veTicks :: [n]

Points in the target interval that should be indicated to the user,

veAxes :: [(n, n)]

Start and end points of axes

veGridLines :: [(n, n)]

Start and end points of grid lines

veAxisLabels :: [(String, n)]

Points in the target interval that should be labelled

veLegend :: [Shape s n]

Additional data

backgroundShapes :: VisualElements s n -> Shape String n

foregroundShapes :: VisualElements String (x, y) -> Shape String (x, y)

Scales that generate visual elements

class ChartVisuals c a b where

The class of scales that have metadata in the form of VisualElements.

Minimal complete definition

visuals

Methods

visuals :: c -> a -> VisualElements String b

newtype Vis a

Constructors

Vis a

Modifiers for visuals

noGrid :: ScaleOptions f a b -> ScaleOptions (NoGrid f) a b

data NoGrid (f :: * -> *) a

noVisuals :: ScaleOptions f a b -> ScaleOptions (NoVisuals f) a b

data NoVisuals (f :: * -> *) a

cartesian :: Extent d -> Extent b -> FontSize String -> LabelOffset (b, d) -> ScaleOptions f a b -> ScaleOptions g c d -> ScaleOptions (Cartesian f g) (a, c) (b, d)

data Cartesian (f :: * -> *) (g :: * -> *) a

Constructors

data family ScaleOptions f a b :: *

Additional parameters for the scale (other than target range).